requirements.txt
Django==1.7.3
django-input-mask==1.3.7.1
